backend/requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==4.2.24
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-jwt-5==1.13.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
slim==0.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0